refactor(calculator): Unify materials section UI
